fix: refactored for forward references, fix use of null values in filters
